requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
apscheduler==3.10.4
python-dotenv==1.0.1
pyyaml==6.0.1
fake-useragent==1.4.0
playwright==1.58.0
